server/folketsserver/requirements.txt
asgiref==3.8.1
django-ratelimit==4.1.0
django-cors-headers==4.3.1
dnspython==2.6.1
Faker==25.0.1
gunicorn==22.0.0
jsonpointer==2.1
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
setuptools==68.2.2
six==1.16.0
wheel==0.41.2